REFACTOR(keen) generation time calculation in listeners v2.0.0#2308.2
The calculation of the generation times in the listeners EvolutionSummary, EvolutionPrinter and EvolutionRecord was updated from using their local properties to directly using the durations in whole milliseconds from GenerationRecord. This change makes the codebase more maintainable by reducing redundancy and potential points of failure by directly utilizing GenerationRecord's duration property. The code was also modified to adhere to the Kotlin coding conventions in EvolutionPrinter and a suppressed warning was added in AbstractEvolutionListener.