FEAT(keen) Add phenotype record and update generation statistics calculation v2.0.0#2308.5
Added a new file "GeneticMaterialRecords.kt" creating a new data class "PhenotypeRecord" which stores the genotype and associated fitness value. This is used to modify various listener classes such as "EvolutionPrinter", "EvolutionSummary", etc.. Now, the generation statistics (best, worst, average fitness, etc.) are calculated based on the sorted population in each generation, which is represented by the phenotype records. This is more accurate and logical as these statistics should reflect the genotype which is subjected to the genetic algorithm. Consequently, the "GenerationRecord" class is also modified to store the population as a list of phenotype records, representing the result of the genetic algorithm in each generation.