REPO Update gradle and plugin versions v2.0.0#2308.10
The gradle version has been updated from 8.2.1 to 8.3-rc-3 for better performance and fixes. Additionally, the versions for plugins have also been updated, including dokka to "1.8.20", ktlint to "11.5.1", kotlinx-coroutines-core to "1.7.3", and slf4j-simple to "2.0.7". The changes aim to utilize better and more efficient functionalities from the newer versions of the plugins. Some of the dependencies have been rearranged too for better organization.