Add RandomSelectorBenchmark and Mutator tests, enhanced JSON serializer
This commit includes the addition of the RandomSelectorBenchmark that evolves a string of characters to match a target string using a genetic algorithm, and further test cases to assess the behavior of the RandomMutator classes. Additionally, a few low-level adjustments have been made, for example the probability of a gene being mutated is now taken into consideration during mutation. Modifications have been made to the JSON evolution serializer such as to automatically handle directory creation for the provided file path, as well as enable the number of spaces used for indentation to be specified. Other files with deviations from the code style guidelines have been reformatted for consistency, some additional comments have been made clearer, and the magic number rule in detekt has been disabled because magic numbers were sufficiently explained in the code and not considered a problem.