Add RandomMutator in keen's mutator architecture
Implemented a new RandomMutator class while converting the existing Mutator into an interface for broader extension in the future. This will allow for more varied types of Mutators such as BitFlipMutator, under the same architecture while ensuring the safeguarding of mutation probability range to be between 0.0 and 1.0. Additionally, the copyright information was updated. This will greatly improve the modularity and extensibility of the mutation functionality within the library.