Refactor mutator architecture and update copyright info
This commit introduces a RandomMutator class and adjusts existing Mutator to be an interface, facilitating more extensibility in terms of adding specific types of mutators like the new BitFlipMutator. This approach maintains consistency on the mutation probability being in the range 0.0 - 1.0. Copyright info was also updated with JetBrains acknowledgement in the README.