Added BukinN6 optimization function and updated IntGeneTest unit tests.
This commit introduces a new optimization function, BukinN6, to the project. This function will be used for genetic algorithms optimization examples. In addition, the commit significantly expands unit tests for IntGeneTest to include equality and hash code checks. This increase in code coverage ensures the robustness and correctness of the IntGene implementation.