Remove SelectionBenchmarkKt and WordGuessingProblem, add optimization files
Removed RandomSelectorBenchmark.kt, RouletteWheelSelectorBenchmark.kt and WordGuessingProblem.kt files as the benchmarks are being revamped. The change is aligned with the reevaluation and reorganization of test suites. Added new optimization files (Functions.kt and Runner.kt) for fitness function tests and modified some existing files to resolve issues in defining range in Random and IntGene and Genotype verification.