Add modify functionality to DoubleChromosomeTest
Current commit adds the modification functionality to the DoubleChromosomeTest along with a test case to ensure the correct behavior. Also, a necessary function is added to the Arbs companion object to create a range of double values. The change improves test coverage and provides more variety in the kind of ranges that can be tested.