Remove commented unused test code
The removed files consist of test classes that have all code commented out, which are apparently not being used in the current context. It's unnecessary to leave these classes in the repository since they contain no actionable code.

By removing these files, the project directory will be less cluttered and easier to navigate. If these tests ever need to be retrieved, they can be found in the git history.

The added files contain new test code for the BitFlipMutator. They were not part of this diff, so no description is provided.