Add new chromosome factory tests & refactor EnforcementTest
In this diff, new unit tests have been added to the `ChromosomeScopeTest.kt` to test the addition of double, integer, and program chromosome factories to a GenotypeScope. These additions are important to ensure that necessary functionality is working correctly.

Further, in `EnforcementTest.kt`, we've changed the checkAll function to use a range of 0 to 50 for arb.list generation, ensuring tests run more efficiently. This change was required for ensuring efficient testing.