Refactor Environment class to accept generic types
The Environment class was refactored to accept generic types, allowing more flexibility in creating environments with different data types. The mutable map in Environment was updated to map from Int to T. Similar modifications were made to the Arbitraries, making them consistent with the newly generic Environment. In the testing files, we made sure to specify the type when initializing an Environment. This enhancement was made to increase the modularity and adaptability of the code. Mirror modifications were applied to associated and dependent files.