Add new method and extend functionality for CollectionRequirement
A new class method 'HaveSize' has been added to the CollectionRequirement class which checks if a collection has a specific size passed as an argument. This enhances the functionality of the collection requirements, adding to the existing 'BeEmpty' method which checks if a collection is empty. Tests have been updated to include the new functionality. Additionally, extensive updates were made to the code and tests of Fun and Environment classes to handle a generic type 'T', improving reusability of these classes. New test files were also created for the Fun class.