Refactor PhenotypeRecord to IndividualRecord and Improve Class Documentation
Refactored the name of the PhenotypeRecord class to IndividualRecord to better describe the entities it is expected to hold. This change aims to enhance readability and understanding for other developers or users who interact with our codebase.

In consonance with the renaming, all references to PhenotypeRecord have been updated to IndividualRecord across multiple sections in various files. In addition to this, detailed documentation was added to the IndividualRecord class to give a more comprehensive description of its use and structure.

Also, Arb helpers were added to facilitate testing of the updated GenerationRecord class along with improvements to its TestBed. Other minor changes include updates to the config in detekt.yml and minor refactoring in some sections of the Kotlin files.

It's worthwhile to note that refactoring PhenotypeRecord to IndividualRecord brings the advantage of better clarity and expressiveness to the code, hence improving maintainability.