Add CompositeAltererTest and refactor CompositeAlterer
Created a new test class, CompositeAltererTest, to test the functionality of the CompositeAlterer class. The tests verify the creation of a CompositeAlterer and its application on a population with different generations and mutations.

The CompositeAlterer class was also refactored for simpler and more succinct code. The primary change includes using the fold function in the invoke method to iterate over the alterers, applying them to each population and accumulating the alterations.

These modifications aim to increase the maintainability of the code and ensure the correct functionality of the CompositeAlterer class.