Revise code and comments in various classes for readability
Made revisions in Reduceable, CompositeAlterer, Terminal, and Constant classes. In Reduceable, an unused method was removed and copyright notice was added. In CompositeAlterer and Terminal, the descriptions and usage instructions in comments were significantly improved and an infection was introduced to make the code more efficient. In Constant, reduced redundancies in properties.  The revisions occurred to increase the readability and maintainability of the code.