Refactor file names, filenames references and argument names
Renamed 'AbstractUniformLengthCrossover.kt' to 'AbstractCrossover.kt' and updated respective references to match the new name. This is more meaningful and consistent with the project's convention. Changed argument name 'inGenotypes' to 'parentGenotypes' in crossover function across multiple classes for clarifying whose genotypes we're referring to. Also, added a couple of useful comments and tests.