Add AverageCrossover functionality and update tests
Added a new class AverageCrossover in the crossover directory to perform crossover on numeric genes by averaging the values. This new method can provide enhanced exploration of the search space for continuous optimization problems. Also, made corresponding changes across several test classes to cover the new functionality and updated an access modifier in AbstractCrossover for an abstract function crossoverChromosomes for more flexibility. Besides that, renamed the test file 'UniformLengthCrossoverTest.kt' to 'CrossoverTest.kt' to reflect the inclusion of tests for the new crossover methods.