Refactor deprecated IntGene constructor, update SinglePointCrossover
In the 'IntGene.kt' file, the deprecated IntToInt constructor was removed in an effort to use more idiomatic stdlib ranges. The 'Ranges.kt' file was updated to add a deprecation warning for IntToInt, directing developers to use ClosedRange instead. The 'SinglePointCrossover.kt' was updated to eliminate the need for explicitly defining `probability`, instead, parameters `chromosomeRate` and `exclusivity` were added with default values to allow more flexibility when using SinglePointCrossover; this was reflected in the update to 'OnesCounting.kt'. A new test file 'SinglePointCrossoverTest.kt' was added to test and ensure the functionality of these changes.