Improve documentation and refactor code for clarity
This commit addresses documentation improvements and refactoring for more clarity in three classes - Match.kt, SinglePointCrossover.kt, and Engine.kt.

In Match.kt, changes were made to the copyright license from BSD Zero Clause to 2-Clause BSD. Correspondingly, copyright holder was updated. The class documentation was also improved for better understandability. An operator function was added to allow the internal predicate to be used in a component-like manner.

In SinglePointCrossover.kt, code was optimised by using 'map' function to couple 'first' and 'second' gene sequences.

The primary changes in Engine.kt were around enhancing the documentation for the Engine class. It has been made more illustrative with details on potential configurable components of the evolution process. ArchiveToString function was removed as it might be unnecessary.

Factoring out verbose logic and enhancing documentation aims to make the codebase more succinct, understandable, and maintainable.