Refactor random API usage and update constructor details
The redundant `Dice.probability()` function in `Core.kt` was removed and now direct usage of `Random.nextDouble()` is used in required places. This simplifies the API by removing an unnecessary layer of abstraction. The unused `probability` import statements were also removed. Additionally, corrected the parameter details mentioned in the constructor of `RandomMutator.kt` and `SubtreeCrossover.kt` to represent the accurate information.