Refactor exception handling and managing methods
Enhanced the way exceptions are thrown and handled, optimizing alterer and mutation processes as a result. Extra exceptions handling for range and filter size have been added to ensure the fitness and gene count match. A new file for the 'GenerationCountTest' was created to test the new conditions. String messages related to exceptions have also been updated for easier understanding and debugging for developers. The new exceptions condition will help developers identify exact errors in the process chain.

Add files for testing evolution stage and alterer results

Added files to implement tests for the evolution results and to enhance test cases for alterer results. The aim is to ensure that errors and bugs in the evolution process are caught in time and fixed accordingly.

Remove unused import statement and redundant statements

Removed redundant statements and imported libraries that are not being used. This cleanup process enhances the readability and efficiency of the code.

Refactor methods names and constructors

Method names were refactored for better understanding and constructors were modified to enhance their efficiency. Changes have been made to the Alterer and the Mutator classes.

Add new testing suites

New test files added for ensuring the functionality of new 'SteadyGenerations' and 'EvolutionResult' classes. These classes are crucial components to determine the termination conditions of the GA process.

Restructure project file organization

Code files were moved around into more appropriate folders for a more coherent file structure.