Refactor algorithm tracking in SteadyGeneration and TargetFitness
The SteadyGeneration and TargetFitness classes were refactored to improve the tracking performance and precision of genetic algorithms. An import of 'IndividualRecord' was added to optimise how populations are sorted. Additionally, a new 'computeSteadyGenerations' method was introduced to enhance the tracking of how generations progress. These changes are meant to provide more accurate readings of target fitness and to enhance the recording of generational progression.