Refactor README.md and update Kotlin codes
The name of the project was clarified in README.md and several explanations on the operations of the framework were expanded. This was done to provide more descriptive information and increase understandability for users. Additionally, some small tweaks were done on Kotlin codes to improve their efficiency and readability. This includes the simplification of accessing the value from Genotype, revision of constants and fitness functions, and fine-tuning of engine parameters.

Importantly, the project acknowledged the support from JetBrains which was not previously mentioned in README.md, giving credit where it is due. This helps to set a good social norm in the open-source community.

The commit also contains an unfinished change on EngineFactoryTest.kt and it requires a reviewers' attention.